Linux paths learning project separate README file addition
Now there is a separate file serving as a README, compared to the previous situation, when the file "linux-paths-learning.txt" served in it. Currently the files overlap, but I hope to reorganize the project a bit (so that the content of "linux-paths-learning.txt" be more maintainable), and then they shall not.