Business simulation project save.txt file removal
The file was just an example save file. The save file is meant to be generated by the user at the path they choose, so including a file with the path that I've happened to choose at one time (i.e., the main project directory) in the repository is unnecessary and hindering a reader's understanding of the project.